requirements.txt
# Run the command in cmd assuming you have pip set up
# to install required dependencies:

# pip install -r requirements.txt

pyserial
ttkthemes
matplotlib